+++ NEWS.md
@@ -10,6 +10,7 @@
 * Gracefully adjust motion filtering parameters if they fall above Nyquist at this TR
 * Modify extract ROIs config to avoid input_regex and always generate it internally from postproc stream
 * Add optional header row for postprocessed confounds TSVs, configurable via postprocess YAML and validated during setup
+* Added extensive checks on write/permission issues with directories and files
 
 # BrainGnomes 0.7-5